Organized security type classes into folder. Changed position shares property to int from long.

+++ Asset Management Platform/Position.cs
@@ -11,13 +11,13 @@ namespace Asset_Management_Platform
         private Security securityInfo;
         public Security SecurityInfo { get { return securityInfo; } }
 
-        private long sharesOwned;
-        public long SharesOwned { get { return sharesOwned; } }
+        private int sharesOwned;
+        public int SharesOwned { get { return sharesOwned; } }
 
         public float Value {
             get { return float.Parse((securityInfo.LastPrice * sharesOwned).ToString()); } }
 
-        public Position(Security security, long shares)
+        public Position(Security security, int shares)
         {
             securityInfo = security;
             sharesOwned = shares;
